+++ src/api/Common.ts
@@ -1,15 +1,14 @@
-import { store } from "@/store/store";
-import axios from "axios";
+import { useStore } from "@/store/store";
+import axios, { AxiosResponse, AxiosError } from "axios";
 import { AxiosInstance } from "axios";
 
 export default class Common {
-    _authHeader: { auth: { username: string; password: string } } = { auth: { username: "", password: "" } };
-    _requestParameter: { auth: { username: string; password: string }; params: any };
+    _authHeader!: Promise<{ auth: { username: string; password: string } }>;
     _axios: AxiosInstance;
 
     constructor(endpoint: string) {
-        this._authHeader = { auth: store.state.loginData };
-        this._requestParameter = { ...this._authHeader, params: {} };
+        this._authHeader = this._getLoginData();
+
         const instance = axios.create({
             baseURL: process.env.VUE_APP_API_BASE_URL + endpoint,
             headers: {
@@ -21,11 +20,33 @@ export default class Common {
         this._axios = instance;
     }
 
+    static async getVersion(endpoint: string): Promise<String> {
+        let version = "unavailable";
+
+        const instance = axios.create({
+            baseURL: process.env.VUE_APP_API_BASE_URL + endpoint,
+            headers: {
+                "Accept": "*/*",
+                "Content-Type": "application/json;charset=UTF-8",
+            },
+        });
+
+        await instance
+            .get(`/version`)
+            .then((response: AxiosResponse) => {
+                version = response.data.versionNumber;
+            })
+            .catch((error: AxiosError) => {});
+        return version;
+    }
+
     getAuthHeader() {
         return this._authHeader;
     }
 
-    setAuthHeader(authHeader: { auth: { username: string; password: string } }) {
-        this._authHeader = authHeader;
+    async _getLoginData() {
+        const store = useStore();
+        const auth = await store.getters.loginData;
+        return { auth: auth };
     }
 }