+++ src/api/HyperledgerManagement.ts
@@ -1,0 +1,11 @@
+import Common from "./Common";
+
+export default class HyperledgerManagement extends Common {
+    constructor() {
+        super("/hyperledger-management");
+    }
+
+    static async getVersion(): Promise<String> {
+        return super.getVersion("/hyperledger-management");
+    }
+}