+++ src/router/index.ts
@@ -7,7 +7,9 @@ import CourseFormSuspenseWrapper from "../views/lecturer/CourseFormSuspenseWrapp
 import AccountFormSuspenseWrapper from "../views/admin/AccountFormSuspenseWrapper.vue";
 import Redirect from "../views/errors/403.vue";
 import ProfileWrapper from "../components/profile/Wrapper.vue";
+import Settings from "../views/common/Settings.vue";
 import PageNotFound from "../views/errors/404.vue";
+import WelcomePage from "../views/common/WelcomePageWrapper.vue";
 
 const routerHistory = createWebHistory(process.env.BASE_URL);
 const suffix: string = " | UC4";
@@ -15,6 +17,14 @@ const suffix: string = " | UC4";
 const router = createRouter({
     history: routerHistory,
     routes: [
+        {
+            path: "/welcome",
+            name: "welcome",
+            component: WelcomePage,
+            meta: {
+                title: "Welcome" + suffix,
+            },
+        },
         {
             path: "/login",
             name: "login",
@@ -123,6 +133,14 @@ const router = createRouter({
             component: Redirect,
             meta: {
                 title: "Permission Denied" + suffix,
+            },
+        },
+        {
+            path: "/settings",
+            name: "settings",
+            component: Settings,
+            meta: {
+                title: "Settings" + suffix,
             },
         },
         {