+++ src/views/admin/Home.vue
@@ -1,6 +1,5 @@
 <template>
     <div>
-        <dev-nav-bar />
         <div class="mt-32 text-4xl text-center font-semibold text-gray-900">Available Accounts</div>
         <div class="mt-8 flex justify-center">
             <admin-account-list />
@@ -9,24 +8,27 @@
 </template>
 
 <script lang="ts">
-    import { store } from "../../store/store";
+    import { checkPrivilege } from "../../use/PermissionHelper";
     import { Role } from "../../entities/Role";
-    import DevNavBar from "../../components/dev_components/DevNavBar.vue";
     import AdminAccountList from "../../components/AdminAccountList.vue";
 
     export default {
         name: "AdminHome",
         components: {
-            DevNavBar,
             AdminAccountList,
         },
 
-        beforeRouteEnter(_to: any, _from: any, next: any) {
-            const myRole = store.state.myRole;
-            if (myRole != Role.ADMIN) {
-                return next("/redirect");
+        async beforeRouteEnter(_to: any, _from: any, next: any) {
+            const response = await checkPrivilege(Role.ADMIN);
+
+            if (response.allowed) {
+                return next();
             }
-            return next();
+            if (!response.authenticated) {
+                return next("/login");
+            }
+
+            return next("/redirect");
         },
     };
 </script>