+++ tailwind.config.js
@@ -17,7 +17,17 @@ module.exports = {
                 },
             },
         }),
-        extend: {},
+        extend: {
+            spacing: {
+                "80": "20rem",
+                "108": "27rem",
+                "160": "40rem",
+                "180": "45rem",
+            },
+            minHeight: {
+                "20": "5rem",
+            },
+        },
     },
     variants: {
         opacity: ["responsive", "hover", "focus", "disabled"],